Fixec circular dependency and other minor items
