+++ tardis/energy_input/tests/conftest.py
@@ -1,8 +1,7 @@
 import pytest
 import numpy as np
 
-from tardis.energy_input.util import SphericalVector
-from tardis.energy_input.base import GXPacket
+from tardis.energy_input.util import SphericalVector, GXPacket, GXPacketStatus
 
 
 @pytest.fixture(scope="function")
@@ -17,6 +16,6 @@ def basic_gamma_ray():
         location=SphericalVector(1.0e15, 0.5),
         direction=SphericalVector(1, 0.25),
         energy=1000.0e3,
-        status="InProcess",
+        status=GXPacketStatus.IN_PROCESS,
         shell=1,
     )
